add match-stats json endpoint
